Adding the initial files for build 001.
Info on the files:
kernel.asm is ran by the bootloader, this file calls an external function located in the kernel.

kernel.c is the actual kernel, right now it contains placeholder code that prints a short message to the screen via VGA graphics. I will add more functions in the next build, such as implementing drivers in different files, so that you don't have to have everything in one file.

kernel.img is the fully compiled, assembled, and linked build, for booting.